Move pinsite related classes to rocket package
