+++ app/src/main/java/org/mozilla/focus/fragment/FirstrunFragment.kt
@@ -27,7 +27,6 @@ import org.mozilla.focus.navigation.ScreenNavigator.Screen
 import org.mozilla.focus.telemetry.TelemetryWrapper
 import org.mozilla.focus.utils.DialogUtils
 import org.mozilla.focus.utils.NewFeatureNotice
-import org.mozilla.rocket.banner.OnClickListener
 import org.mozilla.rocket.periodic.FirstLaunchWorker
 import org.mozilla.rocket.periodic.PeriodicReceiver
