+++ app/src/main/java/org/mozilla/focus/home/HomeFragment.java
@@ -94,6 +94,8 @@ import org.mozilla.rocket.content.ContentPortalView;
 import org.mozilla.rocket.content.ContentRepository;
 import org.mozilla.rocket.content.ContentViewModel;
 import org.mozilla.rocket.download.DownloadIndicatorViewModel;
+import org.mozilla.rocket.home.pinsite.PinSiteManager;
+import org.mozilla.rocket.home.pinsite.SharedPreferencePinSiteDelegate;
 import org.mozilla.rocket.nightmode.themed.ThemedImageButton;
 import org.mozilla.rocket.nightmode.themed.ThemedTextView;
 import org.mozilla.rocket.persistance.History.HistoryDatabase;
