+++ app/src/main/java/org/mozilla/focus/widget/CleanBrowsingDataPreference.java
@@ -20,7 +20,7 @@ import org.json.JSONObject;
 import org.mozilla.fileutils.FileUtils;
 import org.mozilla.focus.R;
 import org.mozilla.focus.history.BrowsingHistoryManager;
-import org.mozilla.focus.home.SharedPreferencePinSiteDelegate;
+import org.mozilla.rocket.home.pinsite.SharedPreferencePinSiteDelegate;
 import org.mozilla.focus.telemetry.TelemetryWrapper;
 import org.mozilla.focus.utils.TopSitesUtils;
 import org.mozilla.rocket.component.PrivateSessionNotificationService;
